+++ src/main/java/interfacePanel/panel/KeyTable.java
@@ -131,7 +131,7 @@ public class KeyTable extends JTable {
 
 	public void createTable(GridBagConstraints c, JPanel frame) {
 		c.gridx = 0;
-		c.gridy = 12;
+		c.gridy = 14;
 		table = new JTable();
 		table.setPreferredScrollableViewportSize(new Dimension(400, 100));
 		table.setFillsViewportHeight(true);